Move parse_address library from client_channel to iomgr.
