+++ grpc.gemspec
@@ -161,8 +161,6 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/filters/client_channel/lb_policy_registry.h )
   s.files += %w( src/core/ext/filters/client_channel/local_subchannel_pool.cc )
   s.files += %w( src/core/ext/filters/client_channel/local_subchannel_pool.h )
-  s.files += %w( src/core/ext/filters/client_channel/parse_address.cc )
-  s.files += %w( src/core/ext/filters/client_channel/parse_address.h )
   s.files += %w( src/core/ext/filters/client_channel/proxy_mapper.h )
   s.files += %w( src/core/ext/filters/client_channel/proxy_mapper_registry.cc )
   s.files += %w( src/core/ext/filters/client_channel/proxy_mapper_registry.h )
@@ -658,6 +656,8 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/iomgr/lockfree_event.cc )
   s.files += %w( src/core/lib/iomgr/lockfree_event.h )
   s.files += %w( src/core/lib/iomgr/nameser.h )
+  s.files += %w( src/core/lib/iomgr/parse_address.cc )
+  s.files += %w( src/core/lib/iomgr/parse_address.h )
   s.files += %w( src/core/lib/iomgr/poller/eventmanager_libuv.cc )
   s.files += %w( src/core/lib/iomgr/poller/eventmanager_libuv.h )
   s.files += %w( src/core/lib/iomgr/polling_entity.cc )
